requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
decouple==0.0.7
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg==3.1.9
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
